backend/requirements.txt
fastapi
uvicorn
websockets
sqlalchemy 
databases
sqlmodel
pydantic
python-dotenv
huggingface_hub
